Can't get it to work right.
Cant get removeAll method to work where it removes all of an inputted element. And cant get Modify ArraySortedList class to incorporate an object of comparator type. a. Add an instance variable of type Comparator<T> b. Add another constructor that accepts an instance of Comparator<T> . C. modify the add method that it would perform comparison based on the rules of "compare" if the Comparator instance is not null; otherwise, it would do comparison based on the rule of "CompareTo".